Move `Domainic::Type::Constraint::Parameter`
This moves `Domainic::Type::Constraint::Parameter` to
`Domainic::Type::Constraint::Specification::Parameter`
